[refactor #62] 삼항 연산자를 사용하여 placeDetails가 null인지 확인하고, 이에 따라 값을 할당
- 구글 정보 조회 시 순서 변경(이름으로 조회 업을 경우 주소로 조회)
- NearbyList DTO 내부 클래스로 이동
- 개발 편의성을 위한 구글 조회 PalentID로 결과 호출 시 시각화 가능하도록 수정
