+++ src/main/java/com/wellcome/WellcomeBE/domain/review/GoogleMapInfoController.java
@@ -2,6 +2,7 @@ package com.wellcome.WellcomeBE.domain.review;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
@@ -11,9 +12,9 @@ public class GoogleMapInfoController {
 
     private final GoogleMapInfoService googlePlaceService;
 
-    @GetMapping("/place/details")
-    public Mono<PlaceReviewResponse> getPlaceDetails() {
-        return googlePlaceService.getPlaceDetails("ChIJn5peWMrlYTURS0SwuW8GLpQ");
+    @GetMapping("/{place}/details")
+    public Mono<PlaceReviewResponse> getPlaceDetails(@PathVariable String place) {
+        return googlePlaceService.getPlaceDetails(place);
     }
 
     @GetMapping("/place/id")
